# docs/swagger.yaml
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role_id:
        type: string
    required:
    - email
    - name
    - password
    - role_id
    type: object
  dto.CreateWorkOrderRequest:
    properties:
      description:
        type: string
      spareparts:
        items:
          $ref: '#/definitions/dto.WorkOrderSparepartLine'
        minItems: 1
        type: array
      title:
        type: string
    required:
    - spareparts
    - title
    type: object
...
  dto.Message:
    properties:
      message:
        type: string
    type: object
...
  dto.PaginatedUserList:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserListResponse'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.PaginatedWorkOrderList:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.WorkOrderResponse'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role_id:
        type: string
    required:
    - email
    - name
    - role_id
    type: object
  dto.UserListResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_logged_in_at:
        type: string
      name:
        type: string
      phone:
        type: string
      role_id:
        type: string
      role_name:
        type: string
    type: object
...
  dto.WorkOrderResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      spareparts:
        items:
          $ref: '#/definitions/dto.WorkOrderSparepartResponse'
        type: array
      status:
        type: string
      title:
        type: string
      wo_number:
        type: string
    type: object
  dto.WorkOrderSparepartLine:
    properties:
      item_id:
        type: string
      quantity:
        type: integer
    required:
    - item_id
    - quantity
    type: object
  dto.WorkOrderSparepartResponse:
    properties:
      item_id:
        type: string
      item_name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
    type: object
...
  global.Response-dto_Message:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Message'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_PaginatedUserList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.PaginatedUserList'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  global.Response-dto_PaginatedWorkOrderList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.PaginatedWorkOrderList'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_UserListResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.UserListResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  global.Response-dto_WorkOrderResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.WorkOrderResponse'
      error_code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: debidarmawan1998@gmail.com
    name: DeboZero Corp Tech Team
  description: This is an API documentation of Progas WMS
  title: PROGAS WMS Backend API
  version: "1.0"
paths:
...
  /maintenance/work-orders:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_PaginatedWorkOrderList'
      security:
      - Bearer: []
      summary: List work orders
      tags:
      - Maintenance
    post:
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWorkOrderRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_WorkOrderResponse'
      security:
      - Bearer: []
      summary: Create work order
      tags:
      - Maintenance
  /maintenance/work-orders/{id}:
    get:
      parameters:
      - description: Work order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_WorkOrderResponse'
      security:
      - Bearer: []
      summary: Get work order by id
      tags:
      - Maintenance
  /maintenance/work-orders/{id}/complete:
    post:
      description: Deduct spare part stock when completing
      parameters:
      - description: Work order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_WorkOrderResponse'
      security:
      - Bearer: []
      summary: Complete work order
      tags:
      - Maintenance
...
  /users:
    get:
      consumes:
      - application/json
      description: List users with pagination and search (Superadmin & Manager only)
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Search by name, email, or phone
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_PaginatedUserList'
      security:
      - Bearer: []
      summary: List users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create user (Superadmin & Manager only)
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Create user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete user (Superadmin & Manager only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get user detail (Superadmin & Manager only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_UserListResponse'
      security:
      - Bearer: []
      summary: Get user by id
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user profile, role, active status, or password (Superadmin
        & Manager only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/global.Response-dto_Message'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - User
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"